More code review feedback.
