+++ daw-app/src/test/java/com/benesquivelmusic/daw/app/ui/StatusStripNoLabelSetTextTest.java
@@ -6,10 +6,8 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -38,7 +36,7 @@ import static org.assertj.core.api.Assertions.assertThat;
  * StatusStripCellSkin.java}. Each is asserted to exist and to be scanned (a guard
  * against a vacuously-empty scan from a broken path). To stay faithful to a
  * SOURCE-level scan it strips {@code //} and {@code /* *\/} comments before
- * matching — exactly as {@code RunLaterConsolidationTest} does — so a Javadoc
+ * matching — via the shared {@link SourceScanSupport} harness — so a Javadoc
  * {@code {@code Label.setText}} mention does not false-match. It then looks for a
  * dotted {@code .setText(} invocation (a receiver-qualified text mutation); the
  * dot requirement means a method <em>named</em> {@code setText} declared in the
@@ -56,19 +54,6 @@ final class StatusStripNoLabelSetTextTest {
     private static final Pattern DOTTED_SET_TEXT =
             Pattern.compile("\\.\\s*setText\\s*\\(");
 
-    /**
-     * One Java comment or one string / text-block literal. Alternation order
-     * matters: a literal is matched before {@code //} and {@code /*} so a
-     * delimiter inside a string is not mistaken for a comment, and a quote inside
-     * a comment is consumed as part of the comment. Mirrors
-     * {@code RunLaterConsolidationTest#COMMENT_OR_STRING}.
-     */
-    private static final Pattern COMMENT_OR_STRING = Pattern.compile(
-            "\"\"\"[\\s\\S]*?\"\"\""        // text block
-            + "|\"(?:\\\\.|[^\"\\\\])*\""   // string literal
-            + "|//[^\\n]*"                  // line comment
-            + "|/\\*[\\s\\S]*?\\*/");       // block comment
-
     /** The four source files that make up the strip and its cell control. */
     private static final List<String> STRIP_SOURCE_FILES = List.of(
             "SessionStatusStrip.java",
@@ -78,7 +63,7 @@ final class StatusStripNoLabelSetTextTest {
 
     @Test
     void stripSourcesDriveTextOnlyViaBindingNeverLabelSetText() throws IOException {
-        Path statusDir = locateDawAppModule()
+        Path statusDir = SourceScanSupport.locateDawAppModule()
                 .resolve("src/main/java/com/benesquivelmusic/daw/app/ui/status");
         assertThat(Files.isDirectory(statusDir))
                 .as("the Session Status Strip sources must live under %s", statusDir)
@@ -98,7 +83,8 @@ final class StatusStripNoLabelSetTextTest {
             // not false-match) but keep string literals — a setText written inside
             // a string literal would still be code worth flagging, and none of
             // these files legitimately contains the substring in a string.
-            String code = stripComments(Files.readString(file, StandardCharsets.UTF_8));
+            String code = SourceScanSupport.stripComments(
+                    Files.readString(file, StandardCharsets.UTF_8));
             if (DOTTED_SET_TEXT.matcher(code).find()) {
                 offenders.add(fileName
                         + "  — contains a dotted .setText(...) call; the strip must "
@@ -120,59 +106,5 @@ final class StatusStripNoLabelSetTextTest {
                         + "never call Label/Labeled.setText(...). Offending files:%n  %s",
                         String.join("\n  ", offenders))
                 .isEmpty();
-    }
-
-    /**
-     * Locate the {@code daw-app} module root. Surefire normally sets the working
-     * directory to the module itself; the fallbacks cover invocations from the
-     * repo root (e.g. {@code mvn -pl daw-app test}). Mirrors
-     * {@code RunLaterConsolidationTest#locateDawAppModule()}.
-     */
-    private static Path locateDawAppModule() {
-        Path cwd = Paths.get("").toAbsolutePath();
-        if (isDawAppModule(cwd)) {
-            return cwd;
-        }
-        Path child = cwd.resolve("daw-app");
-        if (isDawAppModule(child)) {
-            return child;
-        }
-        Path candidate = cwd.getParent();
-        for (int i = 0; i < 5 && candidate != null; i++) {
-            if (isDawAppModule(candidate)) {
-                return candidate;
-            }
-            Path nested = candidate.resolve("daw-app");
-            if (isDawAppModule(nested)) {
-                return nested;
-            }
-            candidate = candidate.getParent();
-        }
-        return cwd;
-    }
-
-    private static boolean isDawAppModule(Path dir) {
-        return Files.isRegularFile(dir.resolve("pom.xml"))
-                && Files.isDirectory(
-                        dir.resolve("src/main/java/com/benesquivelmusic/daw/app"));
-    }
-
-    /**
-     * Removes {@code //} and {@code /* *\/} comments while preserving string /
-     * text-block literals. Mirrors {@code RunLaterConsolidationTest#stripComments}.
-     */
-    private static String stripComments(String source) {
-        Matcher m = COMMENT_OR_STRING.matcher(source);
-        StringBuilder out = new StringBuilder(source.length());
-        while (m.find()) {
-            String token = m.group();
-            // A string / text block (starts with a quote) is code — keep it
-            // verbatim; anything else the alternation matched is a comment and is
-            // replaced with a single space.
-            m.appendReplacement(out, token.charAt(0) == '"'
-                    ? Matcher.quoteReplacement(token) : " ");
-        }
-        m.appendTail(out);
-        return out.toString();
     }
 }
