+++ daw-app/src/test/java/com/benesquivelmusic/daw/app/ui/icons/IconAuditTest.java
@@ -1,5 +1,6 @@
 package com.benesquivelmusic.daw.app.ui.icons;
 
+import com.benesquivelmusic.daw.app.ui.SourceScanSupport;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -7,7 +8,6 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
@@ -76,7 +76,7 @@ class IconAuditTest {
 
     @Test
     void uiSourcesMustNotReferenceLegacyIconFamilies() throws IOException {
-        Path moduleRoot = locateDawAppModule();
+        Path moduleRoot = SourceScanSupport.locateDawAppModule();
         Path srcRoot = moduleRoot.resolve("src/main");
         assertThat(Files.isDirectory(srcRoot))
                 .as("UI sources must live under " + srcRoot)
@@ -181,40 +181,5 @@ class IconAuditTest {
         String trimmed = out.toString().trim();
         if (trimmed.startsWith("*")) return "";
         return out.toString();
-    }
-
-    /**
-     * Locate the daw-app module root. Surefire normally sets the working
-     * directory to the module itself; as a fallback we also check for a
-     * {@code daw-app} child directory (covers invocations from the repo
-     * root such as {@code mvn -pl daw-app test}).
-     */
-    private static Path locateDawAppModule() {
-        Path cwd = Paths.get("").toAbsolutePath();
-
-        // Fast path: Surefire runs from the module dir.
-        if (isDawAppModule(cwd)) return cwd;
-
-        // Fallback: invoked from the repo root — look for daw-app child.
-        Path child = cwd.resolve("daw-app");
-        if (isDawAppModule(child)) return child;
-
-        // Walk up (covers nested invocations or mono-repo layouts).
-        Path candidate = cwd.getParent();
-        for (int i = 0; i < 5 && candidate != null; i++) {
-            if (isDawAppModule(candidate)) return candidate;
-            Path nested = candidate.resolve("daw-app");
-            if (isDawAppModule(nested)) return nested;
-            candidate = candidate.getParent();
-        }
-
-        // Last resort — return cwd and let the assertion below fail with
-        // a clear message.
-        return cwd;
-    }
-
-    private static boolean isDawAppModule(Path dir) {
-        return Files.isRegularFile(dir.resolve("pom.xml"))
-                && Files.isDirectory(dir.resolve("src/main/java/com/benesquivelmusic/daw/app"));
     }
 }
