+++ daw-app/src/test/java/com/benesquivelmusic/daw/app/ui/plugin/NoClassNameFieldAnywhereTest.java
@@ -1,5 +1,6 @@
 package com.benesquivelmusic.daw.app.ui.plugin;
 
+import com.benesquivelmusic.daw.app.ui.SourceScanSupport;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -7,14 +8,11 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -39,8 +37,8 @@ import static org.assertj.core.api.Assertions.assertThat;
  * {@code …app.ui}) plus every {@code .java} under {@code …app.ui.plugin} in
  * {@code src/main} — this test itself lives under {@code src/test}, so the walk
  * is self-excluding. To stay faithful to a SOURCE-level scan it strips
- * {@code //} and {@code /* *\/} comments before matching — exactly as
- * {@code RunLaterConsolidationTest} does — so a historical mention in Javadoc
+ * {@code //} and {@code /* *\/} comments before matching — via the shared
+ * {@link SourceScanSupport} harness — so a historical mention in Javadoc
  * would not false-match. The {@code TextField} assertion is deliberately scoped
  * to exactly the two install-flow files rather than the whole package:
  * {@code PluginBrowser.java} legitimately owns a {@code TextField} — its §6.7
@@ -64,22 +62,9 @@ final class NoClassNameFieldAnywhereTest {
             "PluginBrowser.java",
             "PluginJarScanner.java");
 
-    /**
-     * One Java comment or one string / text-block literal. Alternation order
-     * matters: a literal is matched before {@code //} and {@code /*} so a
-     * delimiter inside a string is not mistaken for a comment, and a quote inside
-     * a comment is consumed as part of the comment. Mirrors
-     * {@code RunLaterConsolidationTest#COMMENT_OR_STRING}.
-     */
-    private static final Pattern COMMENT_OR_STRING = Pattern.compile(
-            "\"\"\"[\\s\\S]*?\"\"\""        // text block
-            + "|\"(?:\\\\.|[^\"\\\\])*\""   // string literal
-            + "|//[^\\n]*"                  // line comment
-            + "|/\\*[\\s\\S]*?\\*/");       // block comment
-
     @Test
     void installFlowSourcesCarryNoClassNameFieldOrTextField() throws IOException {
-        Path uiDir = locateDawAppModule()
+        Path uiDir = SourceScanSupport.locateDawAppModule()
                 .resolve("src/main/java/com/benesquivelmusic/daw/app/ui");
         Path pluginDir = uiDir.resolve("plugin");
         Path managerDialog = uiDir.resolve("PluginManagerDialog.java");
@@ -113,7 +98,8 @@ final class NoClassNameFieldAnywhereTest {
             // Strip comments so a historical Javadoc mention does not
             // false-match; string literals are kept — neither token legitimately
             // appears inside a string in any in-scope file.
-            String code = stripComments(Files.readString(file, StandardCharsets.UTF_8));
+            String code = SourceScanSupport.stripComments(
+                    Files.readString(file, StandardCharsets.UTF_8));
             if (code.contains("classNameField")) {
                 classNameFieldOffenders.add(name
                         + "  — references the retired classNameField (story 304 "
@@ -147,59 +133,5 @@ final class NoClassNameFieldAnywhereTest {
                         + "TextField-free. Offending files:%n  %s",
                         String.join("\n  ", textFieldOffenders))
                 .isEmpty();
-    }
-
-    /**
-     * Locate the {@code daw-app} module root. Surefire normally sets the working
-     * directory to the module itself; the fallbacks cover invocations from the
-     * repo root (e.g. {@code mvn -pl daw-app test}). Mirrors
-     * {@code RunLaterConsolidationTest#locateDawAppModule()}.
-     */
-    private static Path locateDawAppModule() {
-        Path cwd = Paths.get("").toAbsolutePath();
-        if (isDawAppModule(cwd)) {
-            return cwd;
-        }
-        Path child = cwd.resolve("daw-app");
-        if (isDawAppModule(child)) {
-            return child;
-        }
-        Path candidate = cwd.getParent();
-        for (int i = 0; i < 5 && candidate != null; i++) {
-            if (isDawAppModule(candidate)) {
-                return candidate;
-            }
-            Path nested = candidate.resolve("daw-app");
-            if (isDawAppModule(nested)) {
-                return nested;
-            }
-            candidate = candidate.getParent();
-        }
-        return cwd;
-    }
-
-    private static boolean isDawAppModule(Path dir) {
-        return Files.isRegularFile(dir.resolve("pom.xml"))
-                && Files.isDirectory(
-                        dir.resolve("src/main/java/com/benesquivelmusic/daw/app"));
-    }
-
-    /**
-     * Removes {@code //} and {@code /* *\/} comments while preserving string /
-     * text-block literals. Mirrors {@code RunLaterConsolidationTest#stripComments}.
-     */
-    private static String stripComments(String source) {
-        Matcher m = COMMENT_OR_STRING.matcher(source);
-        StringBuilder out = new StringBuilder(source.length());
-        while (m.find()) {
-            String token = m.group();
-            // A string / text block (starts with a quote) is code — keep it
-            // verbatim; anything else the alternation matched is a comment and is
-            // replaced with a single space.
-            m.appendReplacement(out, token.charAt(0) == '"'
-                    ? Matcher.quoteReplacement(token) : " ");
-        }
-        m.appendTail(out);
-        return out.toString();
     }
 }
